Include doctest tests that were not run before
